ci: add K8S and Harness boilerplate code (vars WIP)
In this commit, we have added boilerplate code for K8S and Harness to our continuous integration (CI) pipeline. The variables are still work in progress and will be updated in a future commit.

With these changes, we can now easily deploy our application to Kubernetes clusters and manage it using Harness. This will help streamline our development and deployment process, and ensure that our application is running smoothly in production.